Remove some minor irritants (#363)
* Reduce noise in logs from absent stack file

The operator relies on the automation API to load a stack config file
(e.g., `Pulumi.dev.yaml`) from the working directory. If it fails to
load, the operator assumes it's because there's no such file and
therefore no prior config, logs the error returned by the automation
API, and continues.

It's useful to log the error, because it might be something _other_ than
the file being absent, and failures later might be less
mysterious. However, there's (apparently) a quirk of the logger, which
means if you use the key "Cause" or "Error", with an error value, it'll
add a stack trace. This is pretty annoying, because it doesn't tell you
anything extra, and makes it harder to eyeball the logs.

Therefore: stringify the error, so it can't report a stack trace.

* Recognise YAML runtime during setup

This just adds the YAML runtime to those for which the operator doesn't
need to do any special dependency handling, thereby removing a
misleading log message.

* Avoid saving status when deletion in progress

There's no point to trying to save the status when the object is being
deleted; it will just fail, meaning a needless log message.

* Mention noise reduction in changelog

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>