Use tini as PID 1
The operator runs as PID 1, which is expected to reap zombie processes;
since it doesn't, they get left around to take up room. This commit
installs `tini` and uses it as PID 1, instead.

This removes the use of ssh-agent in the `build/bin/entrypoint` script,
which serves only to prevent go-git from complaining about not finding
an ssh-agent socket -- and doesn't help with authentication (if no SSH
key is supplied, it will simply fail the SSH handshake).

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>