Fail if SSH is used but no private key given
When using SSH, a key must be obtained from somewhere. On the command
line, git would either use the ssh-agent socket, or try to use a key in
~/.ssh. go-git mirrors this, by resorting to ssh-agent if it is not
given any other choices. But in the operator container, it doesn't make
sense too try to use ssh-agent, because there's no chance to add keys to
it -- its only purpose would be to stop go-git from complaining.

So: treat it as an error if someone uses an SSH git URL, but doesn't
supply a private SSH key.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>