Make trivial project for skip-deps test case
We want to be able to check that using `.spec.skipInstallDependencies`
will result in no (e.g.,) `npm install` being run; but, we also need to
see the Stack succeed. So: here is a minimal project, which will succeed
on its own; and, we'll add a spurious extra package, and observe that it
doesn't also get installed.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>