Add GitHub workflows to build, test and distribute the Helm chart
This commit introduces a series of GitHub workflows to the codebase, which will be used to automate various tasks related to continuous integration and deployment. These workflows are defined in the `.github/workflows` directory, and include the following:

- `ct-linting-and-testing.yaml`: This workflow is triggered on pull requests, and is responsible for linting and testing the chart to ensure that it is in a releasable state.

- `chart-publish.yaml`: This workflow is triggered on the main branch only when changes are detected in the chart directory, and is responsible for releasing the chart as traditional bundle and OCI artefact.

- Diverse linter configuration files to tune the linting process to our needs.

Overall, these workflows will help to streamline the development process of the Helm charts and ensure that templating code is always in a deployable state.

Signed-off-by: Engin Diri <engin.diri@ediri.de>