Fail faster when a stack reaches the wrong state
The waitForStack* funcs check that a stack reaches the given state; but,
they will keep checking until the timeout if the stack reaches a state
that's _not_ the one desired, so many tests take until the timeout to fail.

Instead, poll until any state is reached, _then_ check that it's the
right state. This means you have to be a careful to reset the baseline
time, especially if you check for one state, change something, then
check for another. (I had missed that in a couple of places.)

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>