Use stable working directory paths
Previously, the operator asked the OS to create a fresh tmp directory
for each stack run. This is a reasonable enough design, since it means
builds are always from a fresh clone, can't interfere with one another,
and can be safely deleted afterwards. But it also meant that the go
build cache would treat every source file as distinct source code, and
grow its build cache accordingly -- to the tune of tens or hundreds of
megabytes a build.

This commit changes the working directory to a stable path (still under
/tmp), named for the stack object. Since each stack object is uniquely
named and is processed serially, there should be no collisions. (There's
also no security benefit to having random directory names, since a
program or user will ill intent can always just poke around in the
filesystem to find what they want).

This also centralises the creation of the directory, which removes a bit
of repetition.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>