Test the controller works with the Go build cache
A simple test to check that the controller doesn't cause the go build
cache to grow, unduly. Essentially:

 1. Run the stack
 2. Don't change anything
 3. Run the stack again
 4. Check that the Go build cache didn't change between 1. and 3.

If a stack is built in a fresh tmp directory each time (as it was
previously), this test will fail, because at least the stack's main.go
will look like a new file to the cache.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>