Cache Go dependencies
Update the setup-go actions to V3, and use its cache feature.

This will help builds in general; I'm also hoping it will help with the
test involving test/testdata/go-build, which can time out because it
spends a minute or so downloading the (minimal) dependencies for the Go
program therein.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>