Guard against the env being changed under us
When there's tests running concurrently, it's possible that something
will change the environment -- notably, $GOCACHE -- so, this commit adds
a check that the cache dir is the same before and after the test has
run. This is not bullet-proof, but is simple and has a decent chance of
catching errors of logic in the tests.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>
Co-authored-by: Ramon Quitales <ramon@pulumi.com>