Use stable working directory paths (#407)
* Use stable working directory paths

Previously, the operator asked the OS to create a fresh tmp directory
for each stack run. This is a reasonable enough design, since it means
builds are always from a fresh clone, can't interfere with one another,
and can be safely deleted afterwards. But it also meant that the go
build cache would treat every source file as distinct source code, and
grow its build cache accordingly -- to the tune of tens or hundreds of
megabytes a build.

This commit changes the working directory to a stable path (still under
/tmp), named for the stack object. Since each stack object is uniquely
named and is processed serially, there should be no collisions. (There's
also no security benefit to having random directory names, since a
program or user will ill intent can always just poke around in the
filesystem to find what they want).

This also centralises the creation of the directory, which removes a bit
of repetition.

* Test the controller works with the Go build cache

A simple test to check that the controller doesn't cause the go build
cache to grow, unduly. Essentially:

 1. Run the stack
 2. Don't change anything
 3. Run the stack again
 4. Check that the Go build cache didn't change between 1. and 3.

If a stack is built in a fresh tmp directory each time (as it was
previously), this test will fail, because at least the stack's main.go
will look like a new file to the cache.

* Cache Go dependencies

Update the setup-go actions to V3, and use its cache feature.

This will help builds in general; I'm also hoping it will help with the
test involving test/testdata/go-build, which can time out because it
spends a minute or so downloading the (minimal) dependencies for the Go
program therein.

* Let waitForStack* test helpers timeout be supplied

Most of the time we want to fail quite fast if a stack fails to run, but
some stacks -- the one involving Go -- just take a while to run. So: let
a timeout be supplied.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>

* Add changelog entry

---------

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>
Co-authored-by: Ramon Quitales <ramon@pulumi.com>