Use file names to check go build cache stability (#416)
In CI, the test that checks whether building at a stable path works with
the go build cache occasionally fails. The test checks the output of
`du` (disk usage) before and after, and fails (when it fails) with the
difference of a small number of blocks.

Since the test is to check whether new objects are added to the cache,
it's not necessary to check the size. On the assumption that the change
in block usage is incidental, this commit changes that test so it
compares the (sorted, recursive) list of files before and after. New
files would indicate new objects being introduced, and the test will
fail (and if the change to stable build paths is reverted, that's what
happens).