Set terminationGracePeriodSeconds for HelmCharts (#520)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

We would like to override value of terminationGracePeriodSeconds
parameter. Unfortunately pulumi kubernetes operator helm chart is not
setting this value.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

By default, the terminationGracePeriodSeconds parameter is not set, so
the operator will take no action. If this parameter is set then the
value will be overwritten.

impact/no-changelog-required

---------

Co-authored-by: Wojciech Lewandowski <wlk@kmd.dk>