Handle existing workspace directories better (#552)
### Proposed changes

I've changed the stack reconciliation code to clean up existing
workspace directories when it finds them instead of treating them as a
lock and failing forever. We've seen numerous times where the operator
leaves a workspace directory behind for some unknown reason and it
causes the stack to fail to reconcile forever. The only way to resolve
the issue is to remove the directory manually or restart the entire
operator pod.

The operator shouldn't treat directories as locks in my opinion. Given
that they're processed by 1 thread at a time and Pulumi has its own
state lock files (for SaaS & self-hosted backends), an existing
directory shouldn't block the reconciliation and cause a failure that
never resolves itself.
 
Also, I fixed a few typos. Let me know if there's anything I've missed
for this or if there's any concerns with this approach.