[v2] E2E test groundwork (#665)
This adds the groundwork for e2e testing the v2 branch. Try it with
`make test-e2e`.

I haven't enabled an assertion for the random-yaml stack because I
wasn't able to see it reliably reconcile. I was seeing GitHub rate limit
errors as well as errors from the operator when updating status on the
stack.