[v2] Git source (#658)
Implements the git source for v2.

We currently only support secret refs from the Stack API because other
options (fs, env) don't make sense to share between agent and operator.
Auth information is passed via env vars to the agent's fetch container.

The operator also needs access to these secrets in order to pull for the
latest commit. It uses an in-memory `ls-remote` and compares remote
refs against's the Stack's ref. We support long and short forms (e.g.
`refs/heads/foo` or `foo`).

The agent now uses automation API to clone the repo, since it's a
natural place for dogfooding. If the init container is restarted, or uses
a persistent volume for the checkout, there's an edge case where
we won't fast-forward the existing checkout. See
https://github.com/pulumi/pulumi/issues/17288.

Also adds a `shallow` clone option to the stack API for controlling
shallowness. We don't currently perform shallow clones so this is
backwards compatible.

Fixes #655