[v2] Use an apply configuration for the workspace template (#713)
This adds `applyconfiguration-gen` to generate apply configurations for
our APIs. The workspace template is updated to use it.

It's a moderately uphill battle to get apply configurations working with
controllers due to various pieces not including built-in support for
them. In practice this means we need to cut a couple small corners -- in
particular implementing our own `DeepCopy` and tweaking the generated
`GroupVersion` code.

Fixes https://github.com/pulumi/pulumi-kubernetes-operator/issues/704.