[v2] Set config all at once (#718)
We currently issue one `SetAllConfig` RPC for each user-specified
config. This is slow but it has important correctness guarantees:

1. The order we apply config matters -- if the user specifies `foo: foo`
followed by `foo: bar`, the net result must always be `foo: bar`.
2. The Pulumi CLI (and therefore Automation API) only allows specifying
`--path` on an all-or-nothing basis. This is bad for us because we
potentially have a blend of path and non-path keys.

Ideally we would be able to supply all of our configs to the automation
API in a single call, and in the case where _all_ of our config keys are
path-like (or all are not path-like) we actually can do that because we
no longer have limitation (2).

This PR makes that possible in the general case by transforming our
config keys in a way that allows us to treat them as if they are all
path-like.

In particular:
* The agent's `SetAllConfig` handler is modified to take a list of
configs instead of a map in order to preserve config order. The
top-level `path` param is also removed and handled on a per-key basis.
* While resolving configs, we escape any non-path keys so subsequent
path parsing treats them as verbatim. For example `foo.bar` gets escaped
as `["foo.bar"]`.
* We can then supply all of our keys at once to Automation API with
`Path: true`.
* If there are no configs to set then the operator doesn't invoke
`SetAllConfig`.

Fixes https://github.com/pulumi/pulumi-kubernetes-operator/issues/650