Ensure cleanup of Stack in foreground deletion (#760)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Problem

There's two edge cases where a Stack could become stuck in a deleting
state, and this PR addresses both.

1. A missing `Update` object. The stack controller processes the
'current update' (if any) before doing anything else. The stack keeps
the name of the current update in its status block, and performs a
lookup of the named Update object to check its progress. If the update
doesn't exist, it assumes a cache miss and simply requeues. Let's
imagine that the update was actually deleted, e.g. via cascading
foreground delete. In that case, the stack freezes up.
2. A missing `Workspace` object. The stack controller strictly handles
workspace setup before kicking off an update, and doesn't re-create the
workspace if it were to disappear while the update is in progress. If
the workspace goes missing, e.g. due to foreground delete, then the
Update blocks indefinitely and likewise does the Stack.

An additional constraint is that completed `Update` objects serve as a
useful historical record, and would ideally not be garbage-collected
when the associated `Workspace` is deleted. It is desirable that a user
be able to delete an idle workspace, and one would not want to discard
the history in that case.

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

1. Teach the stack controller to place a "stack" finalizer to the
"current" `Update` object, to ensure that it doesn't disappear while a
Stack is waiting for its completion. Once complete, the finalizer is
removed and the completed Update object is then targetable by a cleanup
routine.
2. Be sure to save the status (which removes the finalizer as mentioned
above) in the "fast path" deletion case.
3. Automatically abort an Update when it is deleted.
4. Automatically abort an Update when its managing workspace is deleted.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #753