feat: add `WorkspaceReclaimPolicy` to Stack spec (#804)
### Proposed Changes  

This PR introduces the `WorkspaceReclaimPolicy` option in the Stack
spec. By default, it is set to `Retain`, preserving the current behavior
where a Workspace (along with its associated StatefulSet and pod) is not
deleted after a successful sync. When set to `Delete`, the Workspace pod
is removed after a successful sync, effectively scaling the workload
down to zero.

#### Testing  

- Added envtests to verify that the Workspace object is deleted.  
- Manually validated on a GKE cluster.  

### Related Issues (Optional)  

Fixes #694