Garbage collection for Update objects (#810)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->
Implements a retention policy for completed Update objects, based on a
simple TTL with the default value of 24h. Update objects that are in a
completed state and are older than the TTL are summarily deleted.

#### New API: `Update.spec.ttlAfterCompleted` 

A new field to enable garbage collection of this update, once completed.
Optional; no default value.

#### Update Controller

Updated to implement the cleanup logic. When a TTL is set on a completed
update, the controller looks to the last transition time of the
Completed status condition. If there's time remaining, the controller
requeues the object for later, otherwise it deletes the update object.

#### Stack Controller

Updated to assign a 24-hour TTL on the update objects that it creates. A
future enhancement would be to provide an `updateTemplate` field to
allow for further control. A user may edit an existing Update to change
the TTL.

#### Update Controller Tests

A new unit test was added to exercise the retention logic.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #733