Fix linting issues (#813)
### Proposed Changes

This PR ensures consistency in the list of linters used across
`pulumi/pulumi` and `pulumi/pulumi-kubernetes`. Additionally,
`golangci-lint` is upgraded to `v1.64.2` to resolve an issue where it
was hanging. The codebase has also been updated to address linting
issues identified by the new configuration.

To ensure that linting is upheld going forwards, this PR also enables
linting in CI.

The commits within this PR are structured for incremental review.

### Related Issues (Optional)

Closes: #737