Use audience-scoped access token (#816)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

This PR changes how the operator authenticates to the agent's RPC
endpoint. Before this PR, the operator would pass (as a bearer token)
its own service account token (the contents of
`/var/run/secrets/kubernetes.io/serviceaccount/token`) to the agent. The
problem is that a malicious agent could capture that token, and replay
it to impersonate the operator in other calls to the API server.

The solution implemented in this PR is for the operator to use an
audience-specific token, by setting the `aud` claim of the token to a
workspace-specific value, and by removing the API Server as an audience.
The token for a workspace named `default/random-yaml` now resembles:

```json
{
  "aud": [
    "random-yaml.default"
  ],
  "iss": "https://kubernetes.default.svc.cluster.local",
  "sub": "system:serviceaccount:pulumi-kubernetes-operator:controller-manager"
}
```

The [TokenRequest
API](https://kubernetes.io/docs/reference/kubernetes-api/authentication-resources/token-request-v1/)
is the basis for this solution. It produces short-lived tokens
representing a given service account and intended for an arbitrary
audience(s), signed by the API Server. The operator makes a token
request for its own service account, whenever it needs to access a given
workspace, and then attaches the resultant token to outgoing requests.

The "audience" value is arbitrary, often an https: URL. This PR settles
on the simple form `<name>.<namespace>` representing a given workspace.
e.g. `random-yaml.default`. If we prefer to use a more literal service
URL, it could be `grpc://random-yaml-workspace.default:50051` or
similar. Note how the service endpoint has a `-workspace` suffix.

#### Agent: Audience Configuration and Validation

The agent accepts a new command-line flag `--kube-audience` accepting an
array of values. It is a required parameter when `--auth-mode=kube`.

Incoming tokens are validated to have at least one matching audience.
The agent returns a status code of
[UNAUTHENTICATED](https://grpc.io/docs/guides/status-codes/) when the
audience is mismatched.

#### Operator: Workspace Spec Changes

The operator configures the workspace pod to have a workspace-specific
audience (via the `--kube-audience` flag). Note that the format of the
audience value is encapsulated in the operator.

#### Operator: Connection Manager and Token Cache

It is important to cache the tokens across reconciliation passes. Since
multiple controllers connect to the workspace pod, the controller
manager now provides a shared connection manager to encapsulate this.

The connection manager maintains an LRU cache of _token sources_ (keyed
by workspace). When a controller requests a connection to a given
workspace, the manager vends a token source to produce a series of fresh
tokens. The manager periodically (every 5 minutes) scrubs the cache of
token sources that haven't been used in at least 2 hours, with the
rationale that tokens generally have a 1-hour expiration.

Each token source maintains a cache of the latest token, and rotates the
token when it is within 5 minutes of its expiration time.

#### Operator: New environment variable: POD_SA_NAME

The operator must know its own service account name, to be able to call
into the TokenRequest API. The deployment scripts are adjusted to set
the `POD_SA_NAME` environment variable using the Downward API.

When running the operator in an IDE, one should (at minimum) deploy the
operator's SA (`pulumi-kubernetes-operator/controller-manager`). The
user's own credentials are then naturally used to impersonate the SA,
and there's no longer a need to have a local token file.

#### Operator: RBAC changes

The operator needs new permissions to be able to call the TokenRequest
API to impersonate its own service account. A new `Role` and
corresponding `RoleBinding` is introduced for this purpose. Note that
the permission is not cluster-wide and is scoped to a specific
`resourceName`.

### Testing

New/updated unit tests: 
- connection_test.go
- auth_test.go

### TODOs

- [x] Verify that the RBAC permissions of the operator's service account
do allow for calls to the TokenRequest API.
- [x] Write unit tests for new behavior

### Examples

Here's a full example token:

```json
{
  "aud": [
    "random-yaml.default"
  ],
  "exp": 1739306657,
  "iat": 1739303057,
  "iss": "https://kubernetes.default.svc.cluster.local",
  "jti": "190a47cf-8963-44b4-b77c-18ece58bc4af",
  "kubernetes.io": {
    "namespace": "pulumi-kubernetes-operator",
    "serviceaccount": {
      "name": "controller-manager",
      "uid": "f48ad504-695a-4775-ba52-d66cd7ffdbb8"
    }
  },
  "nbf": 1739303057,
  "sub": "system:serviceaccount:pulumi-kubernetes-operator:controller-manager"
}
```

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #715