Use a prebuilt image for testing helm installation (#822)
### Proposed changes

This PR builds the operator image with a custom tag and loads it into
the KinD cluster to ensure that the Helm Charts can deploy current
versions of the Operator.

A configuration change was also made to ensure we always lint/test all
charts, not just when changes are detected.

### Related issues (optional)

Closes: #739