feat: add CI workflows to sync operator images to ECR and GHCR (#830)
### Proposed changes

This adds a new CI workflow which will sync a pulumi-kubernetes-operator
image on Docker Hub to GHCR and ECR. This workflow is adapted from the
sync workflows in https://github.com/pulumi/pulumi-docker-containers.
The sync workflow is not currently hooked up to the release workflow's
publishing step to enable us to test and validate that syncing actually
works first. Hooking this up will be done in a follow-up PR.

### Related issues (optional)

Related: #811