New example: pulumi-ts (#843)
Contributes a new example showing how to use Pulumi IAC to manage
`Stacks`.

A unique aspect of this example is that it automatically provisions a
Pulumi Cloud access token for the stack to use. Note this important
limitation of `@pulumi/pulumiservice`:

> While you can use this provider to provision access tokens, you’ll
still need to have an access token available to generate an access token
with the provider.