Deprecate ResourceSelectorType "FS" and "Env" (#920)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

This PR makes a breaking API change for security reasons, it removes
support for "FS" and "Env"-type resource selectors in the `Stack` spec.
They allow a Stack to draw environment variable values and files out of
the operator's own pod. This could be used to elevate a user's
privileges by obtaining the operator's own local files, such as its
service account token. in`/var/secrets/...`.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #876